requirements.txt
dnspython
requests
confusables